My attempt at getting the website up and running on Rails 3.2 so that we can use bootstrap and other advanced features.
Tomorrow I will be making sure MetaSearch is working and then I will start migrating files over from the main repo to this one.

I didn't think the time vs. effort payoff would be there for trying to upgrade to 3.2 manually from 3.0.9, so that's why this isn't even a branch.